moved component to folder
