+++ src/components/file-input/UploadFile.vue
